API + UI funciona POST

+++ MagicVilla_API/Modelo/Dto/VillaDto.cs
@@ -11,9 +11,9 @@ namespace MagicVilla_API.Modelo.Dto
         public string Nombre { get; set; }
         public string Detalle { get; set; }
         [Required]
-        public double Tarifa { get; set; }
+        public float Tarifa { get; set; }
         public int Ocupantes { get; set; }
-        public Double MetrosCuadrados { get; set; }
+        public float MetrosCuadrados { get; set; }
         public string ImagenUrl { get; set; }
         public string Amenidad { get; set; }
     }
